+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  * Refactor result handling logic to improve clarity and maintainability
  * Improve the testing of the time package
  * Support `enum` type in result handling
+ * Support custom `ResultHandler` for flexible result processing
 
 ## 1.0.10